requirements.txt
# Para encontrar las ruedas CPU-only de PyTorch
--extra-index-url https://download.pytorch.org/whl/cpu

# Core
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
python-multipart>=0.0.6
openai-whisper==20250625
ffmpeg-python>=0.2.0
jinja2>=3.1.6
aiofiles>=23.1.0
faster-whisper>=0.1.0

# PyTorch CPU-only
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
